auth-service/requirements.txt

fastapi
uvicorn
PyJWT
passlib[bcrypt]
pymysql
cryptography
python-multipart
